refactor: relocate ModalComponent and Home component to improve project structure
