+++ src/components/Layout/Layout.jsx
@@ -1,13 +1,11 @@
 import { Outlet, useLocation } from "react-router-dom";
 import { AnimatePresence, motion } from "framer-motion";
+import { FilterProvider } from "../../context/FilterContext";
 import { useModal } from "../../context/ModalContext";
-import ModalComponent from "../ModalComponent/ModalComponent";
-import Header from "../Header/Header";
+import { ModalComponent, Header } from "./../index";
 import { animationConfig } from "../../utils/animationConfig";
 import { SContainer } from "../../styles/SharedStyles";
 import { SLayout, SMain } from "./Layout.styled";
-import { FilterProvider } from "../../context/FilterContext";
-
 
 const Layout = () => {
   const location = useLocation();
